requirements.txt
python-telegram-bot==20.6
httpx==0.25.2
python-dotenv==1.0.1
nest_asyncio==1.6.0
